+++ src/components/cc-pricing-page/cc-pricing-page.smart.js
@@ -1,10 +1,10 @@
 import './cc-pricing-page.js';
 import '../cc-smart-container/cc-smart-container.js';
 import { fetchAllZones } from '../../lib/api-helpers.js';
+import { defineSmartComponentWithObservables } from '../../lib/define-smart-component-with-observables.js';
 import { fromCustomEvent, LastPromise, unsubscribeWithSignal } from '../../lib/observables.js';
-import { defineComponent } from '../../lib/smart-manager.js';
 
-defineComponent({
+defineSmartComponentWithObservables({
   selector: 'cc-pricing-page',
   params: {
     currency: { type: Object },