+++ CHANGELOG.md
@@ -40,6 +40,7 @@ title: Changelog
 * Upgrade from `@custom-elements-manifest/analyzer@0.4.1` to `@custom-elements-manifest/analyzer@0.6.4`.
 * Fix Custom Element Manifest generation web dev server plugin: disable caching
 * Add JSDoc based typechecking with TypeScript's CLI (just utils.js for now)
+* Rollback the smart-manager to the old low level API and move the observable API to a different module
 
 ## 9.0.0 (2022-07-19)